commit-remove-angular-actualization

+++ back-end/Controllers/Admin/DeletUsar.js
@@ -50,7 +50,23 @@ const deleteUser = async (req, res) => {
                 }
             }
         );
+        const urlAnalitcs = 'http://127.0.0.1:5900/analytics';
 
+        const response = await axios.post(urlAnalitcs, {
+            type: 'delete',
+            name: id_collaborator,
+            id_collaborator: user.name,
+            wage: user.wage,
+            position: user.position
+        }, {
+            headers: {
+                'X-API-Key': process.env.X_API_key,  // A chave da API
+            }
+        });
+        // Verificar a resposta da API de Analytics
+        if (response.status !== 200) {
+            return res.status(500).json({ error: 'Falha ao registrar dados de analytics', details: response.data });
+        }
         // Deletar o usuário da tabela User
         await User.destroy({ where: { id_collaborator } });
